==> 3_make_name_df.ipynb <==
structure_path = "raw_data/alphafold_structures"

# Returns list of (uniprot_id, file_name) e.g., [("P49723", "P49723.pdb"), ...]
def get_uniprot_id_and_file_name_tuple_in_path(path):
    from os import listdir
    return {file.split(".")[0]: file
            for file in listdir(path)}  # "raw_data/alphafold_structures"

feature_names = ["#Gene_Name", "#Uniprot_ID"]

def make_prot_feat_df(gene_seq_path, prot_seq_path, annotation_path):
    import pandas as pd
    import numpy as np
    from Bio import SeqIO
    from Bio.Seq import Seq
    from Bio.SeqRecord import SeqRecord
    from Bio.SeqUtils.ProtParam import ProteinAnalysis
    
    identifiers = list()
    uniprot_ids = dict()
    all_feat_dict = dict()
    
    # Load annotation file
    df = pd.read_csv(annotation_path, sep="\t")
    pdb_id_and_file_names = get_uniprot_id_and_file_name_tuple_in_path(structure_path)
    
    print("Working...")
    for seq_record, prot_record in zip(SeqIO.parse(gene_seq_path, "fasta"), SeqIO.parse(prot_seq_path, "fasta")):
        identifier = seq_record.name
        uniprot_id = df[df["#Gene_Name"]==identifier]["#Uniprot_ID"].values[0]
        
        if uniprot_id == "-" or uniprot_id not in pdb_id_and_file_names.keys():
            continue
            
        identifiers.append(identifier)
        uniprot_ids[identifier] = uniprot_id
    
    for identifier in identifiers:
        attribute_list = list()
        attribute_list.append(identifier)
        attribute_list.append(uniprot_ids[identifier])
        
        all_feat_dict[identifier] = attribute_list
        
    df = pd.DataFrame.from_dict(all_feat_dict, orient='index').reset_index(drop=True)
    
    print(df)
    
    df.columns = feature_names
    print("Done.")

ess_df = make_prot_feat_df("processed_ess/degseq-e.dat", "processed_ess/degaa-e.dat", "processed_ess/degannotation-e-uniprot.dat")
ness_df = make_prot_feat_df("processed_ness/degseq-ne.dat", "processed_ness/degaa-ne.dat", "processed_ness/degannotation-ne-uniprot.dat")
ess_df["label"] = 1
ness_df["label"] = 0

import pickle

ess_df.to_csv("processed_data/yeast_ess_feat_tsv.tsv", sep="\t", index=False)
ness_df.to_csv("processed_data/yeast_ness_feat_tsv.tsv", sep="\t", index=False)

pickle.dump(ess_df, open("processed_data/yeast_ess_feat_df.pickle", "wb"))
pickle.dump(ness_df, open("processed_data/yeast_ness_feat_df.pickle", "wb"))
